+++ src/components/ExcursionInput/durationTimePicker.js
@@ -3,7 +3,7 @@ import TextField from '@material-ui/core/TextField';
 import 'date-fns';
 
 
-import styles from './excursion_input.css';
+import styles from './excursionTextInput.css';
 
 // handleChange - function that will be called when value changes
 // label - value that will be passed to handleChange function
