+++ src/services/excursion_model.js
